:+1: Add view components
We will use these components to render UI content in the constructed layout.

Components responsibility is to render the UI content to the buffer or window
so these only expose setters but getters for its internal state.

Components check if the internal state is changed from previous render and does
nothing if nothing had changed. Additionally, the `render` method returns `true`
if the component is actually rendered so that caller can know if `redraw` is needed.